Added call to destructor in SmartPointer.

+++ GTSL/GTSL/Core.h
@@ -33,5 +33,5 @@ namespace GTSL
 	[[nodiscard]] constexpr typename RemoveReference<T>::type&& MakeTransferReference(T&& arg) noexcept { return static_cast<typename RemoveReference<T>::type&&>(arg); }
 
 	template<typename T, typename R>
-	uint32 RangeForIndex(const T& e, const R& range) { return &e - range.begin(); }
+	int64 RangeForIndex(const T& e, const R& range) { return &e - range.begin(); }
 }
